+++ radix-engine-common/src/data/manifest/custom_extension.rs
@@ -1,7 +1,4 @@
-use super::*;
-use crate::data::scrypto::*;
-use crate::*;
-use sbor::rust::prelude::*;
+use crate::internal_prelude::*;
 
 #[derive(Debug, Clone, PartialEq, Eq, Copy)]
 pub enum ManifestCustomExtension {}
