+++ radix-engine-common/src/data/scrypto/custom_formatting.rs
@@ -1,10 +1,4 @@
-use super::*;
-use crate::address::Bech32Encoder;
-use crate::*;
-use sbor::representations::*;
-use sbor::rust::prelude::*;
-use sbor::traversal::*;
-use utils::ContextualDisplay;
+use crate::internal_prelude::*;
 
 #[derive(Clone, Copy, Debug, Default)]
 pub struct ScryptoValueDisplayContext<'a> {
