+++ radix-engine-common/src/data/scrypto/custom_validation.rs
@@ -1,8 +1,4 @@
-use super::*;
-use crate::*;
-use sbor::rust::prelude::*;
-use sbor::traversal::TerminalValueRef;
-use sbor::*;
+use crate::internal_prelude::*;
 
 //=======================================================================================================
 // NOTE:
