+++ radix-engine/src/vm/vm.rs
@@ -5,10 +5,10 @@ use crate::system::system::KeyValueEntrySubstate;
 use crate::system::system_callback::{SystemConfig, SystemLockData};
 use crate::system::system_callback_api::SystemCallbackObject;
 use crate::types::*;
-use crate::vm::vm::api::ClientApi;
 use crate::vm::wasm::WasmEngine;
 use crate::vm::{NativeVm, ScryptoVm};
 use radix_engine_interface::api::field_lock_api::LockFlags;
+use radix_engine_interface::api::ClientApi;
 use radix_engine_interface::blueprints::package::*;
 
 pub struct Vm<'g, W: WasmEngine> {
