+++ scrypto/src/prelude/mod.rs
@@ -24,13 +24,9 @@ pub use num_traits::{
     sign::Signed,
 };
 pub use radix_engine_interface::blueprints::resource::*;
-pub use radix_engine_interface::constants::*;
 pub use radix_engine_interface::crypto::*;
-pub use radix_engine_interface::data::manifest::model::*;
-pub use radix_engine_interface::data::manifest::*;
-pub use radix_engine_interface::data::scrypto::model::*;
-pub use radix_engine_interface::data::scrypto::*;
 pub use radix_engine_interface::math::*;
+pub use radix_engine_interface::prelude::*;
 pub use radix_engine_interface::time::*;
 pub use radix_engine_interface::traits::*;
 pub use radix_engine_interface::types::*;
